generator: default generated CIA to local directory

+++ generator/generator.py
@@ -282,8 +282,12 @@ class Generator():
         return 0
 
     def makeuniqueid(self):
-        with open("id.json", "r") as f:
-            idlist = json.load(f)
+        try:
+            f = open("id.json", "r")
+        except FileNotFoundError:
+            self.uniqueid = 0xFF400
+            return 0
+        idlist = json.load(f)
         # we are going to use the 0xFF400-0xFF7FF range
         if not idlist:
             self.uniqueid = 0xFF400
@@ -295,17 +299,26 @@ class Generator():
         return 0
 
     def makecia(self):
-        makeromarg = f"{self.cmdarg}makerom -f cia -target t -exefslogo -rsf data/build-cia.rsf -elf data/forwarder.elf -banner data/banner.bin -icon data/output.smdh -DAPP_ROMFS=romfs -major 1 -minor 5 -micro 0 -DAPP_VERSION_MAJOR=1 -o {self.output}"
-        makeromarg += f'-DAPP_PRODUCT_CODE=CTR-H-{self.gamecode} -DAPP_TITLE="{self.title["eng"][0]}" -DAPP_UNIQUE_ID=0x{self.uniqueid}'
+        makeromarg = f'{self.cmdarg}makerom -f cia -target t -exefslogo -rsf data/build-cia.rsf -elf data/forwarder.elf -banner data/banner.bin -icon data/output.smdh -DAPP_ROMFS=romfs -major 1 -minor 5 -micro 0 -DAPP_VERSION_MAJOR=1 -o "{self.output}" '
+        makeromarg += f'-DAPP_PRODUCT_CODE=CTR-H-{self.gamecode} -DAPP_TITLE="{self.title["eng"][0]}" -DAPP_UNIQUE_ID={self.uniqueid}'
+        print(makeromarg)
         makeromrun = subprocess.run(makeromarg, shell=True, capture_output=True, universal_newlines=True)
         if makeromrun.returncode != 0:
             self.message(f"{makeromrun.stdout}\n{makeromrun.stderr}")
             exit()
-        with open("id.json", "r") as f:
-            idlist = json.load(f)
-        idlist.append(self.uniqueid)
-        with open("id.json", "w") as f:
-            json.dump(idlist, f)
+        if not os.path.exists("id.json"):
+            with open("id.json", "w") as f:
+                idlist = [self.uniqueid]
+                json.dump(idlist, f)
+                return 0
+        else:
+            with open("id.json", "r") as f:
+                idlist = json.load(f)
+            if not idlist:
+                idlist = []
+            idlist.append(self.uniqueid)
+            with open("id.json", "w") as f:
+                json.dump(idlist, f)
         return 0
 
     def start(self):
@@ -316,7 +329,7 @@ class Generator():
             self.message("Custom path is not provided. Using path for input file.")
             self.path = self.getrompath(os.path.abspath(self.infile))
         if not self.output:
-            self.output = f"{self.infile}.cia"
+            self.output = f"{os.path.basename(self.infile)}.cia"
         self.message("Getting gamecode...")
         self.getgamecode()
         self.message("Extracting and resizing icon...")
